backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
pydantic==2.10.3
python-multipart==0.0.20
python-dotenv==1.0.1
user-agents==2.2.0          
requests==2.32.3            